\docType{data}
\name{centexamp}
\alias{centexamp}
\title{Example network to explore centrality}
\format{An igraph network object with 10 vertices and 12 edges}
\description{
  This is an example network created for *A User's Guide to
  Network Analysis in R.* It is used to illustrate the
  different ways to assess centrality in a network.
}
\details{
  These data were created to illustrate the
  different ways to assess centrality in a network.
}
\section{Node characteristics:}{
  \itemize{ \item \code{vertex.names} - Letter vertex label 
    \item \code{x} - x coordinates for plotting
    \item \code{y} - y coordinates for plotting
    }
}



\keyword{datasets}

