\docType{data}
\name{ISdf}
\alias{ISdf}
\title{Web of Science search data}
\format{A dataframe object for use by the Bibliometrix package}
\description{
  This is a dataframe containing the results of a Web of Science
  publication search. It contains information on 1,000 publications appearing
  between 2007 and 2106 in the journal *Implementation Science.*
}
\details{
  These data are stored as a data frame that can be read by the bibliometrix
  R package. Further bibliometric analyses, including network analyses, can
  be done using various bibliometrix analytic functions.
}

\author{
  Douglas Luke
}

\references{
  Dervis, H. (2019). Bibliometrix analysis using Bibliometris: An R package.
  \emph{Journal of Scientometric Research, 8(3),} 156-160.

}


\keyword{datasets}

